Subsume bug fixes & tolerance for isMoreGeneral
Modified getUpperBound and getLowerBound to facilitate providing a
tolerance factor to enlarge range considered when checking if a
classifier can subsume another.

Also added script for generating testing and training data to custom
proportions with even distribution of each classification. This script
is called splitTestAndTrain.py and generates data_test.txt and
data_train.txt in the features/ directory.